chore(*) Bring latest master commits into release/0.35 branch (#997)
* fix(db) do not try to cascade-delete plugins twice (#972)

Looks like code in the latest master inherited modules such as
`kong/db/strategies/cassandra/routes.lua` and
`kong/db/strategies/postegres/routes.lua` from EE 0.34, but that was necessary
because in CE 0.14 `routes` were new-DAO and `plugins` were old-DAO. Now both
are new-DAO and the DAO does cascade delete by itself, so manually
cascade-deleting plugins (+ the custom workspace and RBAC code there)
shouldn't be necessary anymore.

* fix(admin) correctly delete rbac_role_entity when deleting rbac_user - EBB-19 (#982)

* fix(admin) correctly delete rbac_role_entity when deleting rbac_user

* Delete associated rbac_role_entities
* Delete rbac_roles of admins when deleting admins

* test(admins) create a fake cache:get function

* fix(admin) basic auth token is the user supplited password (#985)

* fix(workspaces) support multiple ':' characters in workspaced fields (#960)

* refactor(admin) improve auth response messages (#984)

Fixes EBB-35

* chore(rate-limiting) update locks shared dict

Update shm used for locks in rate-limiting libraries. Previously, as
there were no dedicated locks shared dictionary, the `kong_cache` dict
was used; said dict was renamed a while back, and its usage refined
(it's now used only as db cache).

* fix(db) select_all on global workspace scope

Make `select_all` comply with the current "global workspace behavior" -
all dao methods should run in global scope if the workspace context is
empty.

* tests(ee-integration) add select_all workspace scope tests

Add tests for select_all on single and global scope.

* tests(ee-integration) select_all tests refactor

Make select_all tests run independently from each other.

* fix(files) portal files slash in name (#986)

* fix(files) handle files entities with slash in name

* feat(portal) add helper to add splat route

* feat(portal) call before filter in splat route before if needed

* tests(portal) add/reenable tests for slash in file name

* tests(portal) set openid-connect portal render test to pending

* feat(portal) pass self, db, helpers to original before filter

* refactor(dev creds) move dev cred logic to crud_helpers share with developer route (#990)

* fix(dev creds) refactor credential endpoints to be used between admin and portal api

* tests(dev creds) update integration tests for developer creds

* fix(admins) filter admins by workspace INTF-1374 (#998)

The original consumers_rbac_users_map wasn't workspace-aware,
and admins is essentially a refactor of that table. This approach
is less invasive than making admins workspaceable and still meets
the goal of isolating admins to appropriate workspaces.

* fix(migrations) fix updating username for the bootstrap user INTF-1441 (#979)

Add back the defaults for created_at and updated_at that were in PR #941.
Ensure created_at on the basicauth_credential is the right precision to
match basicauth_credentials schema. (created_at on plugins is in seconds,
not milliseconds.)

* refactor(workspaces) remove redundant helper functions (#994)

* refactor(workspaces) remove redundant helper functions

* fix(DAO) RBAC check for cascade delete

+++ kong/db/dao/init.lua
@@ -19,7 +19,6 @@ local fmt          = string.format
 local deep_copy    = utils.deep_copy
 
 local workspaces = require "kong.workspaces"
-local ws_helper  = require "kong.workspaces.helper"
 local rbac       = require "kong.rbac"
 
 
@@ -193,7 +192,7 @@ local function check_update(self, key, entity, options, name)
   end
 
 
-  ws_helper.remove_ws_prefix(self.schema.name, entity_to_update)
+  workspaces.remove_ws_prefix(self.schema.name, entity_to_update)
 
   local ok, errors = self.schema:validate_update(entity_to_update)
   if not ok then
@@ -424,7 +423,7 @@ local function generate_foreign_key_methods(schema)
 
         local params = { [name] = unique_value }
         local constraints = workspaceable[self.schema.name]
-        ws_helper.apply_unique_per_ws(self.schema.name, params, constraints)
+        workspaces.apply_unique_per_ws(self.schema.name, params, constraints)
 
         local row, err_t = self.strategy:select_by_field(name, params[name], options)
         if err_t then
@@ -432,7 +431,7 @@ local function generate_foreign_key_methods(schema)
         end
 
         if not row then
-          local pk, err_t = ws_helper.resolve_shared_entity_id(self.schema.name,
+          local pk, err_t = workspaces.resolve_shared_entity_id(self.schema.name,
                                                               { [name] = unique_value },
                                                                workspaceable[self.schema.name])
           if err_t then
@@ -477,7 +476,7 @@ local function generate_foreign_key_methods(schema)
         end
 
 
-        local pk, err_t = ws_helper.resolve_shared_entity_id(self.schema.name,
+        local pk, err_t = workspaces.resolve_shared_entity_id(self.schema.name,
           { [name] = unique_value },
           workspaceable[self.schema.name])
         if err_t then
@@ -487,7 +486,7 @@ local function generate_foreign_key_methods(schema)
           return self:update(pk, entity, options)
         end
 
-        ws_helper.remove_ws_prefix(self.schema.name, entity)
+        workspaces.remove_ws_prefix(self.schema.name, entity)
         -- luacheck: ignore
 
         local entity_to_update, rbw_entity, err, err_t = check_update(self, unique_value,
@@ -522,7 +521,7 @@ local function generate_foreign_key_methods(schema)
           validate_options_type(options)
         end
 
-        local pk, err_t = ws_helper.resolve_shared_entity_id(self.schema.name,
+        local pk, err_t = workspaces.resolve_shared_entity_id(self.schema.name,
                                                             { [name] = unique_value },
                                                              workspaceable[self.schema.name])
 
@@ -577,7 +576,7 @@ local function generate_foreign_key_methods(schema)
         local constraints = workspaceable[self.schema.name]
         local params = {[name] = unique_value}
 
-        local workspace = ws_helper.apply_unique_per_ws(self.schema.name, params, constraints)
+        local workspace = workspaces.apply_unique_per_ws(self.schema.name, params, constraints)
 
         local row, err_t = self.strategy:upsert_by_field(name, params[name],
                                                          entity_to_upsert, options)
@@ -616,7 +615,7 @@ local function generate_foreign_key_methods(schema)
       methods["delete_by_" .. name] = function(self, unique_value, options)
         validate_unique_type(unique_value, name, field)
 
-        local pk, err_t = ws_helper.resolve_shared_entity_id(self.schema.name,
+        local pk, err_t = workspaces.resolve_shared_entity_id(self.schema.name,
                                                             { [name] = unique_value },
                                                             workspaceable[self.schema.name])
         if err_t then
@@ -655,6 +654,17 @@ local function generate_foreign_key_methods(schema)
         end
 
         local cascade_entries = find_cascade_delete_entities(self, entity)
+
+        if (not options or not options.skip_rbac) and  -- not skipping rbac
+          (not rbac.validate_entity_operation(entity, self.schema.name) or
+            not rbac.check_cascade(cascade_entries, ngx.ctx.rbac))  then
+          -- operation or cascading not allowed
+          local err_t = self.errors:unauthorized_operation({
+            username = ngx.ctx.rbac.user.name,
+            action = rbac.readable_action(ngx.ctx.rbac.action)
+          })
+          return nil, tostring(err_t), err_t
+        end
 
         local _
         _, err_t = self.strategy:delete_by_field(name, unique_value, options)
@@ -740,7 +750,7 @@ function DAO:select_all(fields, options)
 
   -- prepend workspace prefix
   local tbl = deep_copy(fields)
-  ws_helper.apply_unique_per_ws(schema.name, tbl, constraints)
+  workspaces.apply_unique_per_ws(schema.name, tbl, constraints)
 
   local rows, err_t
 
@@ -752,7 +762,7 @@ function DAO:select_all(fields, options)
 
   -- if no rows are found, attempt finding shared entities
   if #rows == 0 then
-    local pk, err = ws_helper.resolve_shared_entity_id(self.schema.name, fields, constraints)
+    local pk, err = workspaces.resolve_shared_entity_id(self.schema.name, fields, constraints)
     if err then
       return nil, tostring(err_t), err_t
     end
@@ -804,7 +814,7 @@ function DAO:select(primary_key, options)
   local table_name = self.schema.name
   local constraints = workspaceable[table_name]
 
-  local ok, err = ws_helper.validate_pk_exist(table_name, primary_key, constraints)
+  local ok, err = workspaces.validate_pk_exist(table_name, primary_key, constraints)
   if err then
     local err_t = self.errors:database_error(err)
     return nil, tostring(err_t), err_t
@@ -962,7 +972,7 @@ function DAO:insert(entity, options)
     entity_to_insert.cache_key = self:cache_key(entity_to_insert)
   end
 
-  local workspace, err_t = ws_helper.apply_unique_per_ws(self.schema.name, entity_to_insert,
+  local workspace, err_t = workspaces.apply_unique_per_ws(self.schema.name, entity_to_insert,
                                                          workspaceable[self.schema.name])
   if err then
     return nil, tostring(err_t), err_t
@@ -1028,7 +1038,7 @@ function DAO:update(primary_key, entity, options)
   end
 
   local constraints = workspaceable[self.schema.name]
-  local ok, err = ws_helper.validate_pk_exist(self.schema.name,
+  local ok, err = workspaces.validate_pk_exist(self.schema.name,
                                               primary_key, constraints)
   if err then
     local err_t = self.errors:database_error(err)
@@ -1058,7 +1068,7 @@ function DAO:update(primary_key, entity, options)
     end
   end
 
-  ws_helper.apply_unique_per_ws(self.schema.name, entity_to_update, constraints)
+  workspaces.apply_unique_per_ws(self.schema.name, entity_to_update, constraints)
 
   local row, err_t = self.strategy:update(primary_key, entity_to_update, options)
   if not row then
@@ -1077,7 +1087,7 @@ function DAO:update(primary_key, entity, options)
     end
   end
 
-  ws_helper.remove_ws_prefix(self.schema.name, rbw_entity)
+  workspaces.remove_ws_prefix(self.schema.name, rbw_entity)
   self:post_crud_event("update", row, rbw_entity)
 
   return row
@@ -1103,7 +1113,7 @@ function DAO:upsert(primary_key, entity, options)
   -- Check entity is already in workspace, if yes update workspace_entity
   -- relation else add relation
   -- FIXME try upsert to workspace_entity
-  local is_update, err = ws_helper.validate_pk_exist(self.schema.name,
+  local is_update, err = workspaces.validate_pk_exist(self.schema.name,
                                                      primary_key, constraints)
 
 
@@ -1144,7 +1154,7 @@ function DAO:upsert(primary_key, entity, options)
     return nil, tostring(err_t), err_t
   end
 
-  local workspace = ws_helper.apply_unique_per_ws(self.schema.name, entity_to_upsert, constraints)
+  local workspace = workspaces.apply_unique_per_ws(self.schema.name, entity_to_upsert, constraints)
 
   local row, err_t = self.strategy:upsert(primary_key, entity_to_upsert, options)
   if not row then
@@ -1197,7 +1207,7 @@ function DAO:delete(primary_key, options)
   end
 
   local constraints = workspaceable[self.schema.name]
-  local ok, err = ws_helper.validate_pk_exist(self.schema.name,
+  local ok, err = workspaces.validate_pk_exist(self.schema.name,
                                               primary_key, constraints)
   if err then
     local err_t = self.errors:database_error(err)
@@ -1208,7 +1218,7 @@ function DAO:delete(primary_key, options)
     return true
   end
 
-  ws_helper.apply_unique_per_ws(self.schema.name, primary_key, constraints)
+  workspaces.apply_unique_per_ws(self.schema.name, primary_key, constraints)
 
   local ok, errors = self.schema:validate_primary_key(primary_key)
   if not ok then
@@ -1334,7 +1344,7 @@ function DAO:row_to_entity(row, options)
     return nil, tostring(err_t), err_t
   end
 
-  ws_helper.remove_ws_prefix(self.schema.name, entity, options and options.include_ws)
+  workspaces.remove_ws_prefix(self.schema.name, entity, options and options.include_ws)
 
   return entity
 end
